+++ R/count_objects.R
@@ -111,7 +111,7 @@
 #' \donttest{
 #' library(pliman)
 #' ###### deprecated use analyze_objects() instead ############
-#' img <- image_import(image_pliman("soybean_touch.jpg"))
+#' img <- image_pliman("soybean_touch.jpg")
 #' analyze_objects(img)
 #'
 #' }